requirements.txt
aiohttp==3.8.3
aiosignal==1.3.1
asn1crypto==1.5.1
async-timeout==4.0.2
attrs==22.1.0
beautifulsoup4==4.11.1
bip-utils==2.7.0
cbor2==5.4.3
certifi==2022.9.24
cffi==1.15.1
charset-normalizer==2.1.1
coincurve==17.0.0
crcmod==1.7
ecdsa==0.18.0
ed25519-blake2b==1.4
fake-useragent==1.0.1
frozenlist==1.3.3
idna==3.4
multidict==6.0.2
numpy==1.23.5
pretty-utils @ git+https://github.com/SecorD0/pretty-utils@main
py-sr25519-bindings==0.2.0
pycparser==2.21
pycryptodome==3.15.0
PyNaCl==1.5.0
python-dotenv==0.21.0
requests==2.28.1
semantic-version==2.10.0
six==1.16.0
soupsieve==2.3.2.post1
typing_extensions==4.4.0
urllib3==1.26.12
yarl==1.8.1
